Move initial wallet fetching
fetchWallets is now first called in the wallet context provider so that
it runs when the app is first loaded. Previously, it was first called
in the Wallets page. So if a user opened the app and immediately tried
to add a new task before loading the Wallets page, the wallets in
WalletSelect were empty since fetchWallets was not called and the
wallets provided by useContext was not filled.