Change wallet select to handle single choices
When a user is creating a task, the wallet select component should offer
multiple wallet choices and create a task for each wallet. However, when
a user is editing a task, they should only be able to select a single
wallet and update the task wallet. It would not make sense for users to
edit a single wallet into multiple.

This change removes the 'multiple' property in the event that a user is
editing a task rather than creating a task.